+++ src/compiler/command_line_parser/mod.rs
@@ -6,7 +6,9 @@ mod lines_1500_2000;
 pub use lines_1500_2000::*;
 mod lines_2000_2500;
 pub use lines_2000_2500::*;
-mod lines_2500_3000;
-pub use lines_2500_3000::*;
+mod lines_2500_2750;
+pub use lines_2500_2750::*;
+mod lines_2750_3000;
+pub use lines_2750_3000::*;
 mod lines_3000_end;
 pub use lines_3000_end::*;
